fix: enforce strict mode typing across all service logic

+++ src/modules/like/like.service.ts
@@ -3,9 +3,9 @@ import LikeRepository from "../like/like.repository.js";
 import TweetRepository from "../tweet/tweet.repository.js";
 
 export class LikeSerivce {
-  likeRepository: any;
-  tweetRepository: any;
-  commentRepository: any;
+  likeRepository: LikeRepository;
+  tweetRepository: TweetRepository;
+  commentRepository: CommentRepository;
 
   constructor() {
     this.likeRepository = new LikeRepository();
@@ -13,31 +13,31 @@ export class LikeSerivce {
     this.commentRepository = new CommentRepository();
   }
 
-  async toggleLike({ likable, onModel, user }) {
+  async toggleLike({ likable, onModel, user }: any) {
     let likeable;
     if (onModel === "Tweet") {
       likeable = await this.tweetRepository.getTweet(likable);
     } else if (onModel === "Comment") {
       likeable = await this.commentRepository.get(likable);
     }
-    let exists = await this.likeRepository.findby({
+    let exists: any = await this.likeRepository.findby({
       onModel: onModel,
       user: user,
       likable: likable,
     });
-    if (exists) {
+    if (exists && likeable) {
       console.log(exists);
-      likeable.likes.pull(exists.id);
+      (likeable.likes as any).pull(exists.id);
       await likeable.save();
       await this.likeRepository.destroy(exists.id);
-    } else {
+    } else if (likeable) {
       let newLike = await this.likeRepository.create({
         onModel: onModel,
         user: user,
         likable: likable,
       });
       console.log("newLike", newLike);
-      likeable.likes.push(newLike);
+      likeable.likes.push(newLike.id as any);
       await likeable.save();
     }
